pyproject.toml
[project]
name = "openjia"
version = "0.1.0"
description = "OpenJIA agent harness for verifiable long-running software development"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8",
    "python-dotenv>=1.0",
    "click>=8.0",
    "rich>=13.0",
    "pathspec>=0.11",
]

[project.optional-dependencies]
llm = [
    "openai>=1.50",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
openjia = "openjia.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
